+++ Game_Loop_Test2/src/com/company/Objects/StaticObjects/Coord.java
@@ -4,15 +4,15 @@ package com.company.Objects.StaticObjects;
  * Created by Magnus on 25-04-2016.
  */
 public class Coord {
-    public double X = 0;
-    public double Y = 0;
+    public float X = 0;
+    public float Y = 0;
 
-    public Coord(double x, double y){
+    public Coord(float x, float y){
         X = x;
         Y = y;
     }
 
-    public void NewPos(Vector vec, double velocity){
+    public void NewPos(Vector vec, float velocity){
         this.X = this.X + (vec.X * velocity);
         this.Y = this.Y + (vec.Y * velocity);
     }
